Refactored email router and email proxy. Now: email service as well as email proxy use same class to establish a smtp connection
