+++ Backend/app/dependency/email_service.py
@@ -1,8 +1,7 @@
 from fastapi.params import Depends
+from app.email.smtp_conn import SmtpConnection
+from app.service.email_service import EmailService
 from app.dependency.smtp_connection import get_smtp_connection
-from app.email_router.email_service import EmailService
-
-from app.email_router.smtp_conn import SmtpConnection
 
 
 def get_email_service(smtp_connection: SmtpConnection = Depends(get_smtp_connection)):
